astar/requirements.txt
requests>=2.32.0
python-dotenv>=1.0.0
numpy>=1.26.0
matplotlib>=3.8.0
scikit-learn>=1.5.0
joblib>=1.3.0
scipy>=1.11.0
